Refactor eslint rules to its own folder
